ai/requirements.txt
google-generativeai
torch
langchain-community
langchain-huggingface
langchain
langchain-chroma
sentence-transformers
